+++ index.html
@@ -230,6 +230,6 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-  <script src="js\main.js"></script>
+  <script src="js\main.js" type="module"></script>
 </body>
 </html>
